Got something running.  More work to do...
The lexer needs to be worked on.  I've only got identifiers getting recognized.  In addition,
it appears that I won't be able to avoid editing the denada.y file (vs. editing the
EBNF).  So I need to make sure I'm done with the ebnf part of the tool chain and then
switch over to denada.y exclusively (which will then have to be version controlled).