Added support (in rules) for specifying grammar for children
This includes two built-in rules for now,  and .   are the children
of the current rule and it is the default path for a definition if nothing is specified.

The syntax for specifying the child rules is 'rule>childrule'.  If no '>' appears, the
default is 'rule>'.

In the future, I'll need to add a mechanism for storing other rule names, but for now
I just have  and .