Revamped context concept
This version is much improved because it provides complete access ot the rule hierarhcy.
It supports relative paths (and paths in general) as well as the notion of a formal
root context.  It allows looking things up by rule name as well.  Overall, I think it
is conceptually simpler to understand and it requires fewer lines of code to implement.